get triangle example building on all platforms
