+++ tests/triangle/glfw.go
@@ -1,0 +1,63 @@
+//go:build !js
+
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+
+	"github.com/go-gl/glfw/v3.3/glfw"
+	"github.com/rajveermalviya/go-webgpu/wgpuext/glfw"
+)
+
+func main() {
+	if err := glfw.Init(); err != nil {
+		panic(err)
+	}
+	defer glfw.Terminate()
+
+	glfw.WindowHint(glfw.ClientAPI, glfw.NoAPI)
+	window, err := glfw.CreateWindow(640, 480, "go-webgpu with glfw", nil, nil)
+	if err != nil {
+		panic(err)
+	}
+	defer window.Destroy()
+
+	s, err := InitState(window, wgpuext_glfw.GetSurfaceDescriptor(window))
+	if err != nil {
+		panic(err)
+	}
+	defer s.Destroy()
+
+	window.SetKeyCallback(func(w *glfw.Window, key glfw.Key, scancode int, action glfw.Action, mods glfw.ModifierKey) {
+		// Print resource usage on pressing 'R'
+		if key == glfw.KeyR && (action == glfw.Press || action == glfw.Repeat) {
+			report := s.instance.GenerateReport()
+			buf, _ := json.MarshalIndent(report, "", "  ")
+			fmt.Print(string(buf))
+		}
+	})
+
+	window.SetSizeCallback(func(_ *glfw.Window, width, height int) {
+		s.Resize(width, height)
+	})
+
+	for !window.ShouldClose() {
+		glfw.PollEvents()
+
+		err := s.Render()
+		if err != nil {
+			fmt.Println("error occurred while rendering:", err)
+
+			errstr := err.Error()
+			switch {
+			case strings.Contains(errstr, "Surface timed out"): // do nothing
+			case strings.Contains(errstr, "Surface is outdated"): // do nothing
+			case strings.Contains(errstr, "Surface was lost"): // do nothing
+			default:
+				panic(err)
+			}
+		}
+	}
+}
